completed 3.2 courses screen
