+++ src/Kanbas/index.tsx
@@ -3,20 +3,52 @@ import Account from "./Account";
 import Dashboard from "./Dashboard";
 import KanbasNavigation from "./Navigation";
 import Courses from "./Courses";
-import './style.css'; 
+import './style.css';
+import * as db from "./Database";
+import { useState } from "react";
 
 export default function Kanbas() {
+    const [courses, setCourses] = useState<any[]>(db.courses);
+    const [course, setCourse] = useState<any>({
+        _id: "1234", name: "New Course", number: "New Number",
+        startDate: "2023-09-10", endDate: "2023-12-15", description: "New Description",
+    });
+    const addNewCourse = () => {
+        setCourses([...courses, { ...course, _id: new Date().getTime().toString() }]);
+    };
+    const deleteCourse = (courseId: any) => {
+        setCourses(courses.filter((course) => course._id !== courseId));
+    };
+    const updateCourse = () => {
+        setCourses(
+            courses.map((c) => {
+                if (c._id === course._id) {
+                    return course;
+                } else {
+                    return c;
+                }
+            })
+        );
+    };
     return (
         <div id="wd-kanbas" className="kanbas-container">
             <nav>
-            <KanbasNavigation />
+                <KanbasNavigation />
             </nav>
             <div className="kanbas-content wd-main-content-offset p-3">
                 <Routes>
                     <Route path="/" element={<Navigate to="/Kanbas/Dashboard" />} />
                     <Route path="/Account/*" element={<Account />} />
-                    <Route path="/Dashboard" element={<Dashboard />} />
-                    <Route path="/Courses/:cid/*" element={<Courses />} />
+                    <Route path="Dashboard" element={
+                        <Dashboard
+                            courses={courses}
+                            course={course}
+                            setCourse={setCourse}
+                            addNewCourse={addNewCourse}
+                            deleteCourse={deleteCourse}
+                            updateCourse={updateCourse} />
+                    } />
+                    <Route path="/Courses/:cid/*" element={<Courses courses={courses} />} />
                     <Route path="/Calendar" element={<h1>Calendar</h1>} />
                     <Route path="/Inbox" element={<h1>Inbox</h1>} />
                 </Routes>
